feat: add rest endpoint for batch deal import

+++ src/main/java/com/bloomberg/fxdeals/dtos/res/FxDealResDTO.java
@@ -1,11 +1,11 @@
 package com.bloomberg.fxdeals.dtos.res;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record FxDealResDTO(String dealId,
                            String fromCurrency,
                            String toCurrency,
                            LocalDateTime dealTimestamp,
-                           BigInteger dealAmount) {
+                           BigDecimal dealAmount) {
 }
